+++ src/agents/AgentCoordinator.ts
@@ -5,8 +5,6 @@ import {BasicAgent} from './BasicAgent';
 import {UniswapPoolScannerAgent} from './UniswapPoolScannerAgent';
 import {UniswapPriceScannerAgent} from './UniswapPriceScannerAgent';
 import {UniswapVerificationAgent} from './UniswapVerificationAgent';
-import {FeedAgent} from './FeedAgent';
-import {BlockMintingAgent} from './BlockMintingAgent';
 
 @injectable()
 export class AgentCoordinator {
@@ -17,15 +15,11 @@ export class AgentCoordinator {
     @inject(UniswapPoolScannerAgent) UniswapPoolScannerAgent: UniswapPoolScannerAgent,
     @inject(UniswapPriceScannerAgent) UniswapPriceScannerAgent: UniswapPriceScannerAgent,
     @inject(UniswapVerificationAgent) UniswapVerificationAgent: UniswapVerificationAgent,
-    @inject(FeedAgent) FeedAgent: FeedAgent,
-    @inject(BlockMintingAgent) BlockMintingAgent: BlockMintingAgent,
   ) {
     this.agents = {
       UniswapPoolScannerAgent,
       UniswapPriceScannerAgent,
       UniswapVerificationAgent,
-      FeedAgent,
-      BlockMintingAgent,
     };
   }